Setup initial package.json
I've entered myself as the author, not sure if this is ok, or if maybe something like "Rich Harris, Vatroslav Vrbanic" would be better which would kind of imply both of us made it by working closely together, which is not true. Or maybe "Rich Harris (Svelte), Vatroslav Vrbanic (svelte-accmod)"? 'asvelte-ccmod' is just modifying Svelte (like a patch), it IS actually Svelte, but it also feels wrong to leave "Rich Harris" as author and distribute it like that being hosted on my account. I'm considering to ask Rich directly about this before making it public if I don't find any codex concerning this.